Fixed bug in determining if the message should be checked for viruses.
Filtering should not be performed for very large messages. The length of the
message was only being considered for the SpamAssassin tests, and not for the
ClamAV tests. Sending a large message to ClamAV causes a time out, which
results in a 451 status being sent back to Postfix, with the result that the
message cannot get past the content filter.