Enabled the spam filter to continue processing after quarantining a message.
A Postfix after-queue filter is supposed to close the connection to the second
server if it decides that the message should be quarantined. However the first
server can continue to send more messages to it. The previous behaviour was
that the filter closed the connection to the first server on receiving the next
command, with the result that an error message was printed in the mail log and
the next message's delivery was deferred. With this fix, when the next command
is sent to the filter after a message has been quarantined then a new
connection is opened to the second server, the greeting from the second server
is discarded and a synthesised EHLO command is sent to the second server.