No XCLIENT command for remote client addresses in the POP before SMTP table.
If the remote address is in the POP before SMTP table, then the email
originated from a mail client that may send a HELO command that will fail the
access checks on the relayed mail server. If the XCLIENT command is not sent
in such cases then the name of the relaying host will be sent, which will pass
the remote access checks.