Removed the auto-accepting feature of the greylist policy.
This feature would automatically accept new addresses from a domain if a
threshold of successful deliveries from that domain had been passed. This is
problematic because spam can be forged to come from common domains, such as
gmail.com. The default greylist interval was changed some time ago to be five
minutes, so there is no longer a great need for this feature.