Added a blacklist table.
Entries in this table are added by the blacklist policy. The entries are used
by the greylist policy to determine if greylisting needs to be performed,
because if the IP address appears in the blacklist table then the blacklist
policy will have already rejected the message. This is necessary because
Postfix still evaluates the recipient restrictions (where greylisting is
performed) even if the sender restrictions (where blacklisting is performed)
have already rejected the message.