Removed the ability to use multiple databases that was introduced in 3886129.
This was originally introduced to allow blacklisting to query the database on
the main mail server and a backup mail server, but is no longer needed now that
a backup server is no longer used. It was only necessary because the two
servers did not host the same set of domains, and only a subset were shared
between them. If both servers host the same set of domains, and therefore the
backup server acts as a full backup, then there is also no need for one server
to query the other's database. This simplifies the design of the policy class.

This commit also moves the isGreylisted function from greylist.py to mixin.py
to solve an import issue, since it is used by both the BlacklistPolicy and
GreylistPolicy classes.