Major reworking of the SMTP access policy daemons.
There is a major bug in Postfix's support for SMTP access policy delegation
that has been present since at least Postfix 2.3.3 and is still present in the
latest 2.5.6 release. The effect of the bug is that only a single policy
service per smtpd service is reliable. The bug occurs if more than one policy
is defined, and more than a single request is handled by the smtpd service
before the spawn service controlling the policy daemon has exited. In this
case, although the first policy daemon will be sent the correct information
about the current message, subsequent requests to the other policy daemons for
the same message will occasionally be sent a mixture of information about the
current message and the previous message handled by the policy daemon. The
instance, sender and recipient are correct, but all other information is taken
from the other message.

The simple workaround for this is to combine all of the policies into a single
policy daemon, which is what this commit does.