Added a utility to report spam.
This utility expects the spam message to be fed to it on standard input as an
attachment to another message. This allows a simple spam reporting mechanism to
be set up on a special address which feeds any messages it receives to this
utility. A user receiving a spam message in their inbox then simply forwards
that message as an attachment to the special email address.