Added the ability to track mail originating from local users.
This fixes a problem introduced by the hard blacklisting of sites and class C
networks. If somebody sends email from such a site and has not sent email to
a local user before, then their message will be rejected as it will not have a
greylist entry. This commit adds the ability to track correspondence
originating from a local user via both a policy access delegation daemon and
via the spamcheck process. If a local user has already sent email to an address
that originates from a host or network that is blacklisted, then the blacklist
will now be suppressed for that address.