Perform all blacklist tests before determing the blacklist status.
This fixes a bug in previous versions of the blacklist policy, in which the
tests for class C networks were only performed if neither the IP address nor
the host name resulted in a blacklisting. This would prevent a message from
being hard blacklisted due to a class C network block if the email came from an
IP address or host that was currently soft blacklisted. With this commit, all
tests are performed and the resulting status is then determined as being the
most restricitve one.