Improve the creation of spam test descriptions in the database.
This commit removes the previous mechanism, which searched through the
SpamAssassin configuration files to find the descriptions, and instead parses
the response sent by spamc. One complication is that if a test does not have a
description, then spamc will set the description equal to the name of the test.
In this case, a SpamTest object is created with a null description, which can
be updated later if a description is provided for that test.