Fixed an issue with incorrectly encoded email headers.
Some email headers are not encoded according to RFC 2047. In this case we
attempt to convert the header to Unicode assuming that the encoding is utf8,
and if that fails we destructively convert the string to ascii encoding by
dropping all non-ascii characters. This is necessary because all expressions
evaluated by Mako must either be Unicode strings or ascii-encoded strings.