Corrected the transaction isolation level management.
The previous mechanism issued a "set transaction isolation level serializable"
command, which was intended to make every transaction have a serializable
isolation level. But this command only applies to the current transaction, and
subsequent transactions will revert to the PostgreSQL default, which is read
committed isolation. The new mechanism is to issue a "set session
characteristices as transaction isolation level serializable" command. However,
this command only applies to subsequent transactions, and the current
transaction will still be in the default read committed isolation level.
Therefore in order to force all transactions to be in the desired isolation
level, the session is committed immediately after the set session
characteristics command is issued.

This commit also allows the isolation level to be selected, in preparation for
changes to be made to the spamcheck program.