Auto-whitelist scores are now updated if any disabled tests are triggered.
If the message is in one of the character sets for which some tests are
disabled, then the auto-whitelist entry will have been incorrectly increased by
spamd with the scores for the disabled tests, and so the auto-whitelist score
is adjusted by this commit. This also requires that spamcheck be able to
execute its transaction in the read committed isolation level, since spamd will
have updated the auto-whitelist records after the transaction has begun for the
spamcheck program.