Work around an issue in SQLAlchemy.
SQLAlchemy does not gracefully handle the case where one attempts to read a
deferred attribute after the record which the object maps has been deleted
from the database. Instead of throwing an exception it enters an infinite loop.
The fix is to retrieve the value of the attribute into a local variable before
deleting the record.