Fix a bug when an attached message has been encoded.
There is a bug in Python's email module in the way that it processes the payload
of messages with a content type of message/rfc822 if the contents have been
encoded with a content transfer encoding. Instead of decoding the payload
before created the message object, the email module simply creates a message
object with no headers and with the payload set to the encoded contents.