Fixed a bug in the processing of hard class C rejections.
There was a bug in the algorithm for performing a hard rejection of mail from
a blacklisted class C network. If a message delivery was attempted before the
hard threshold was reached, then a soft rejection would be performed and a
greylist record would be created for the delivery attempt. If the hard
threshold for that class C network was reached before the next delivery
attempt after the greylist period expired, then the new blacklisting status was
not taken into account and the message would be accepted instead of being
issued a hard rejection.