Corrected a hole in the updating of the AWL when spam is received.
This commit accounts for spam in which the envelope header is different from
the From: header, in which case SpamAssassin would not have updated the entry
corresponding to the envelope header. If the AWL entry for the envelope header
was below the spam threshold, then this allowed the message to avoid
greylisting or blacklisting in the policy check because it would have passed
the isAccepted method. This commit updates the AWL entry corresponding to the
envelope sender address and the IP address of the remote SMTP server.