Preserve the case of the envelope sender address for spam and viruses.
This commit changes the name of the mail_from column in the spam and viruses
tables to bounce, and no longer converts the sender address into lower case in
these tables. This allows spam to be bounced at a later time to the same
address that was used to send the original message. Although RFC 822 states
that email addresses are case-insensitive some mail servers, such as Google
groups, do not follow this and will not accept a bounce sent to the lower cased
version of the sender address.