Ensure that the original message text is available to email extractors.
This is in preparation for updating the AWL table when a valid DKIM signature
is present in the message. The original message contents will always be
available in the original_message attribute of an EmailExtractor object. It is
necessary to do this because deserialisation of an email message object is not
guaranteed to exactly reproduce the message from which the object was created.