Fixed DKIM verification failures when verifying an embedded message.
DKIM requires that the message being verified has not been changed in any way,
which is not possible to guarantee if the email has been parsed and then
serialised by Python's email module. The algorithm that extracts the original
message from an attachment was therefore updated to guarantee this by bypassing
the email module and extracting the message as a substring of the enclosing
message fed to the spam reporting script.