Allow a spam message to be reported with a text/plain content type.
When forwarding a spam message as an attachment, the content type is set to
message/rfc822. In certain circumstances, such as a mailing list relaying spam,
the message has to be edited before being sent to avoid the mailing list
software being identified as the spammer. When manually attaching an edited
message, Thunderbird sets the content type to text/plain. In order for the
consume script to identify this as the message to be processed, the file name
of the message must be spam.txt if the content type is text/plain.