Ignore DKIM signatures when reporting spam.
The auto_whitelist table will now be updated for all rows that match the sender
address and the IP, irrespective of whether or not that row contains a signedby
value. This allows spammers to be reported who on different occassions have had
their messages both signed and unsigned, so that future spam from them will be
flagged regardless or whether or not it has been signed.