Added a driver shell script for delivery of quarantined messages.
The driver script ensures that the python script will run in the correct
virtualenv environment. The previous version of the delivery script contained
a hack which modified the site module so that the modules imported by the
script could be found, but actually ran the script with the system python
executable. This caused problems if the system python and the python installed
in the virtualenv environment used different UCS settings.