Made whether greylisting should consider auto-whitelisting configurable.
The default behaviour is that if no greylist entry currently exists then mail
will be accepted if the sender is currently on the auto-whitelist. However, if
a spam storm hits the server with multiple messages coming from the same sender
to multiple recpipients and at least one of the recipients has greylisting
disabled and the spam is accepted for that recipient then that will result in
the spam messages being delivered to all of the recipients. The default
behaviour should only be used if greylisting has not been disabled for all
users of the system.