Added processing of honeypot addresses when performing the spam checks.
If any of the recipients is a honeypot address and the sender was not
whitelisted then the message is treated as spam, and the recipients
for the message will be any of the recipients that were not honeypot
addresses. If the sender was whitelisted and there were no
non-honeypot addresses then the message is bounced backed to the
sender, otherwise the message is delivered normally.