Delete all relevant greylist entries when a message is marked as spam.
The previous policy was to only remove the greylist entries
corresponding to the recipients of the message, but for cases
where a spam network is using several servers on the same class C
network this would not prevent a message from being accepted by the
greylist mechanism once a message to another recipient has been
recognised as spam from another server on the same network. The new
policy is to remove all of the greylist records for the class C network
once a message has been identified as spam or a virus.