Workaround for a hard to reproduce issue in Postfix.
There is an issue in Postfix that is very difficult to reproduce. Occasionally
when an after content filter attempts to connect to an SMTP server on localhost
on a given port, if there is an instance of an SMTP server listening on that
port on a different IP address then the connection will be made to that server
instead of the one configured for localhost in the Postfix masters.cf file.
Since the other server was oftern configured to use an after content filter
this leads to an infinite loop which terminates with a too many hops error
message. The workaround is to configure another SMTP server on localhost on a
different port and have the after content filter communicate with that one
instead. Shell configuration files were set up for all of the shell
initialisation scripts to allow arguments to be configured on the Python class
constructors that implement all of the spamfilter services.