Prevent a runtime error if SpamAssassin reports no spam tests.
With the changes to log all messages sent to honeypot addresses it is now
possible for a message to be saved in the spam table even if SpamAssassin
reports the message as being clean. This commit fixes a runtime error that
previously occurred when attempting to enter a message into the spam table when
that message did not have any spam tests associated with it as reported by
SpamAssassin.