Added blacklist mechanism for suspicious connection attempts.
In an attempt to avoid RBL blacklists, some spammers use many servers
in the same class C network, send spam for a few minutes from one of
those servers and then move on to another server, continuing this
process until the block of servers has been exhausted. The servers are
usually ordered by increasing IP address. This commit adds thresholds
to the maximum number of servers from a class C network that can
connect to a Postfix SMTP server within a configurable time
interval before such connections will be rejected. The algorithm does
not currently consider the ordering of the IP addresses of the
connection attempts, but could be enhanced to do this in the future.